make the backends compositor-neutral
The backends used to have lots of code dealing with weston specific configs.
To allow them to be used by other compositors with their own configurations
remove all the usage of weston_config from the backends code and move it
in weston.c. The backends are now initialized with the new
weston_compositor_init_backend() function which takes a pointer to a
weston_backend_config struct. Each backend now installs one header where
a subclass of weston_backend_config is defined, and the compositor creates
an instance of the backend-specific config struct and passes it to
weston_compositor_init_backend().
This means that the compositor now needs to know about all the backends
to be able to instantiate them, but no backend specific API such as xcb
or libdrm leaks through, so there are no new dependencies for the compositor.

A backend may now set a vfunc pointer to create a new compositor. Not all do
though, and use instead a callback function to configure a new output
that comes from below the compositor, instead of above it.