build a libweston.so used by the weston binary
The library is versioned, so a -N is appended to its name, e.g. libweston-1.so.
The same is done for the pc file, the include and lib/weston dirs.
This allows multiple libweston versions to be installed together.
Additionally, a new libweston.pc is introduced in paraller with weston.pc,
which is used when building a Weston plugin.