compositor: add a protocol for sharing buffers with clients
This commit adds a new protocol for sharing a surface's buffer with
a given client. Given a weston_surface and a wl_client, a new resource
is created with interface _wl_shared_buffer, which maps the content
of the current attached buffer on that surface. The _wl_shared_buffer
is sent to the client and can be accessed by it.
Only shm buffers are supported currently, EGL ones need a new EGL
extension.