xwayland: make the plugin usable by libweston compositors
This patch follows a similar approach to the taken to detach the backends
from weston. But instead of passing a configuration struct when loading the
plugin, we use the plugin API registry to register an API, and to get it
in the compositor side.
This API allows to spawn the Xwayland process in the compositor side, and
to deal with signal handling.
A new function is added in compositor.c to load and init the xwayland.so plugin.

Signed-off-by: Giulio Camuffo <giuliocamuffo@gmail.com>